move application to separate app; keep unprotected groups open
