+++ foodsaving/applications/__init__.py
@@ -1,0 +1,8 @@
+from django.apps import AppConfig
+
+
+class ApplicationsConfig(AppConfig):
+    name = 'foodsaving.applications'
+
+    def ready(self):
+        from . import receivers  # noqa: F401
