+++ foodsaving/applications/templates/application_accepted.mjml
