+++ foodsaving/applications/templates/application_declined.mjml
