+++ foodsaving/groups/serializers.py
@@ -5,9 +5,8 @@ from django.utils.translation import ugettext as _
 from rest_framework import serializers
 from rest_framework.exceptions import ValidationError, PermissionDenied
 
-from foodsaving.conversations.models import Conversation
 from foodsaving.groups.models import Group as GroupModel, GroupMembership, Agreement, UserAgreement, \
-    GroupNotificationType, GroupApplication, GroupApplicationStatus
+    GroupNotificationType
 from foodsaving.history.models import History, HistoryTypus
 from foodsaving.history.utils import get_changed_data
 from . import roles
@@ -142,54 +141,6 @@ class GroupDetailSerializer(GroupBaseSerializer):
             payload=self.initial_data
         )
         return group
-
-
-class GroupApplicationSerializer(serializers.ModelSerializer):
-    conversation = serializers.SerializerMethodField()
-
-    class Meta:
-        model = GroupApplication
-        fields = [
-            'id',
-            'user',
-            'group',
-            'conversation',
-            'questions',
-            'answers',
-            'status',
-        ]
-        read_only_fields = [
-            'user',
-            'questions',
-            'status',
-        ]
-
-    def get_conversation(self, application):
-        return Conversation.objects.get_for_target(application).id
-
-    def validate(self, attrs):
-        if GroupApplication.objects.filter(
-            group=attrs.get('group'),
-            user=self.context['request'].user,
-            status=GroupApplicationStatus.PENDING.value,
-        ).exists():
-            raise serializers.ValidationError(_('Application is already pending'))
-        return attrs
-
-    def validate_group(self, group):
-        if group.is_member(self.context['request'].user):
-            raise serializers.ValidationError(_('You are already member of the group'))
-        if group.is_open:
-            raise serializers.ValidationError(_('You cannot apply to open groups'))
-        return group
-
-    def save(self, **kwargs):
-        group = self.validated_data['group']
-        return super().save(
-            **kwargs,
-            user=self.context['request'].user,
-            questions=group.get_application_questions_or_default(),
-        )
 
 
 class AgreementSerializer(serializers.ModelSerializer):
