+++ foodsaving/groups/tasks.py
@@ -1,25 +1,24 @@
 from datetime import timedelta
 
 from anymail.exceptions import AnymailAPIError
-from django.contrib.auth import get_user_model
 from django.db.models import Count
 from django.utils import timezone
 from huey import crontab
-from huey.contrib.djhuey import db_periodic_task, db_task
+from huey.contrib.djhuey import db_periodic_task
 from influxdb_metrics.loader import write_points
 from raven.contrib.django.raven_compat.models import client as sentry_client
 
 from config import settings
-from foodsaving.groups import stats, emails
+from foodsaving.applications.stats import get_group_application_stats
 from foodsaving.groups.emails import (
     prepare_user_inactive_in_group_email,
     prepare_group_summary_data,
-    prepare_new_application_notification_email,
-    prepare_application_accepted_email,
-    prepare_application_declined_email,
+    calculate_group_summary_dates,
+    prepare_group_summary_emails
 )
-from foodsaving.groups.models import Group, GroupStatus, GroupNotificationType
+from foodsaving.groups.models import Group, GroupStatus
 from foodsaving.groups.models import GroupMembership
+from foodsaving.groups.stats import get_group_members_stats, get_group_stores_stats, group_summary_email
 from foodsaving.utils import stats_utils
 
 
@@ -30,9 +29,9 @@ def record_group_stats():
     points = []
 
     for group in Group.objects.all():
-        points.extend(stats.get_group_members_stats(group))
-        points.extend(stats.get_group_stores_stats(group))
-        points.extend(stats.get_group_application_stats(group))
+        points.extend(get_group_members_stats(group))
+        points.extend(get_group_stores_stats(group))
+        points.extend(get_group_application_stats(group))
 
     write_points(points)
 
@@ -70,7 +69,7 @@ def send_summary_emails():
 
     for group in groups:
 
-        from_date, to_date = emails.calculate_group_summary_dates(group)
+        from_date, to_date = calculate_group_summary_dates(group)
 
         if not group.sent_summary_up_to or group.sent_summary_up_to < to_date:
 
@@ -78,7 +77,7 @@ def send_summary_emails():
 
             context = prepare_group_summary_data(group, from_date, to_date)
             if context['has_activity']:
-                for email in emails.prepare_group_summary_emails(group, context):
+                for email in prepare_group_summary_emails(group, context):
                     try:
                         email.send()
                         email_count += 1
@@ -91,7 +90,7 @@ def send_summary_emails():
             group.sent_summary_up_to = to_date
             group.save()
 
-            stats.group_summary_email(
+            group_summary_email(
                 group,
                 email_recipient_count=email_recipient_count,
                 feedback_count=context['feedbacks'].count(),
@@ -116,30 +115,3 @@ def mark_inactive_groups():
         if not group.has_recent_activity():
             group.status = GroupStatus.INACTIVE.value
             group.save()
-
-
-@db_task()
-def notify_members_about_new_application(application):
-    users = application.group.members.filter(
-        groupmembership__in=GroupMembership.objects.active().with_notification_type(
-            GroupNotificationType.NEW_APPLICATION
-        ),
-    ).exclude(
-        groupmembership__user__in=get_user_model().objects.unverified_or_ignored(),
-    )
-
-    for user in users:
-        try:
-            prepare_new_application_notification_email(user, application).send()
-        except AnymailAPIError:
-            sentry_client.captureException()
-
-
-@db_task()
-def notify_about_accepted_application(application):
-    prepare_application_accepted_email(application).send()
-
-
-@db_task()
-def notify_about_declined_application(application):
-    prepare_application_declined_email(application).send()
